ci_template: mark the manifest job as not allowed to fail
With gitlab 13.12 the behavior of jobs that depend of manual
jobs changes, and they are considered skipped by default.

But in our pipeline we want to enforce that the whole pipeline
runs and succeds and the manual status is only there to reduce
CI load.

For more details see:

https://about.gitlab.com/blog/2021/05/20/dag-manual-fix/

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-ci/-/merge_requests/404>